requirements.txt
# Apache Airflow e dependências principais
apache-airflow==2.10.5

# Bibliotecas para integração com provedores (opcional, de acordo com suas necessidades)
apache-airflow-providers-amazon
apache-airflow-providers-google
apache-airflow-providers-docker
apache-airflow-providers-postgres

# Bibliotecas úteis para manipulação de dados
pandas
numpy

# Para chamadas HTTP e integração com APIs
requests

# Visualização de dados (opcional)
matplotlib
seaborn

# Para conexões a bancos de dados
sqlalchemy
psycopg2-binary  # PostgreSQL
mysql-connector-python  # MySQL
